work on day 7
I think I understand the problem better. Also at one point I had the right answer in the variable signal, but the program didn't halt. Could it be a concidence that that number showed up (it was 139629729), or is it a sign that I'm close to getting it correct? or... is it possible that this is an issue with my intcode computer?